Moved common includes for both loader and bpf into the common/ folder
