+++ server/src/tools.mk
@@ -10,7 +10,7 @@ BPFTOOL_SRC_DIR := $(BPFTOOL_DIR)/src
 BPFTOOL_OUT_DIR := $(BIN)/bpftool
 BPFTOOL := $(BPFTOOL_OUT_DIR)/bpftool
 
-INCLUDES := -I$(LIBBPF_OUT_DIR)
+INCLUDES += -I$(LIBBPF_OUT_DIR)
 
 $(LIBBPF): | $(LIBBPF_OUT_DIR)
 	$(MAKE) -C $(LIBBPF_SRC_DIR) BUILD_STATIC_ONLY=1 OBJDIR=$(LIBBPF_OUT_DIR) \
